Cambia la variable x de double a int para que el resultado sea un número entero

+++ src/main/java/ejerciciosbucles/Ej3.java
@@ -6,7 +6,7 @@ public class Ej3 {
 
     
     public static void main(String[] args) {
-       double x = 15;
+       int x = 15;
 
         while (x <= 15*Math.pow(2, 24)) {
             System.out.print(x);
